Enable non-x86 architectures
This charm uses the elasticsearch apt repository, which is currently missing
metdata for non-x86 architectures. 

See: https://github.com/elasticsearch/elasticsearch/issues/9299

The elasticsearch deb, however, is architecture-independent and I've found that it seems to work just fine on at least one non-x86 architecture (arm64). This change overrides the architecture setting for the elasticsearch repository so that all architectures will try to fetch the amd64 version of the metadata, making the same packages visible.